moved some classes for better CA
