+++ src/main/java/adapters/custom_game/custom_game_UI_adapters/CustomGameSubmissionManager.java
@@ -1,0 +1,5 @@
+package adapters.custom_game.custom_game_UI_adapters;
+
+public class CustomGameSubmissionManager {
+
+}
