+++ src/main/java/adapters/custom_game/custom_game_UI_adapters/EditorOnClick.java
@@ -1,4 +1,4 @@
-package user_interface.custom_game.custom_game_panels;
+package adapters.custom_game.custom_game_UI_adapters;
 import adapters.custom_game.custom_game_file_adapters.EditorTile;
 
 import java.awt.event.MouseAdapter;
@@ -7,7 +7,7 @@ import java.awt.event.MouseEvent;
 /**
  * Reacts to clicks on CustomGameEditorPanel
  */
-class EditorOnClick extends MouseAdapter {
+public class EditorOnClick extends MouseAdapter {
     private final EditorTile TILE;
     private int rightClickCount = 0;
