+++ src/main/java/user_interface/custom_game/custom_game_panels/CustomGameEditorPanel.java
@@ -36,6 +36,7 @@ class CustomGameEditorPanel implements ICustomGamePanel {
 
 //        JTextField staminaField = new JTextField("stamina");
 //        editorFrame.add(staminaField, BorderLayout.LINE_END);
+//        or combine with the submission button to use PAGE_END for both
 
         JButton submissionButton = new JButton("submit");
         editorFrame.add(submissionButton, BorderLayout.PAGE_END);
